memoize tailwind class generation
